Abstract into batches.ts
