+++ automkv.ts
@@ -1,6 +1,23 @@
-import {getExe, printUsage} from "./utils.ts";
+import {getExe} from "./utils.ts";
 import Runner from "./runner.ts";
 import Watcher from "./watcher.ts";
+
+function printUsage() {
+    console.log("Usage:");
+    console.log("  automkv watch <file>");
+    console.log("    Monitor the target 'automkv.yml' file's destination folder for any");
+    console.log("    changes, and run the edits in the file when any updates have");
+    console.log("    finished.")
+    console.log("  automkv watch <folder>");
+    console.log("    Monitor the target folder (and subfolders) for files ending in");
+    console.log("    'automkv.yml'.  Any files found will have their destination folders");
+    console.log("    monitored for changes, and the edits in the file will run when any");
+    console.log("    updates have finished.");
+    console.log("  automkv run <target_file>");
+    console.log("    Apply the edits in a 'automkv.yml' file immediately to all target");
+    console.log("    files in the target folder");
+    Deno.exit(0);
+}
 
 const mkvpropedit = await getExe("mkvpropedit");
 const mkvextract = await getExe("mkvextract");
