+++ utils.ts
@@ -1,42 +1,4 @@
-import {log, path, yaml} from "./deps.ts";
-
-import {Batch} from "./types.ts";
-
-export function printUsage() {
-    console.log("Usage:");
-    console.log("  automkv watch <file>");
-    console.log("    Monitor the target 'automkv.yml' file's destination folder for any");
-    console.log("    changes, and run the edits in the file when any updates have");
-    console.log("    finished.")
-    console.log("  automkv watch <folder>");
-    console.log("    Monitor the target folder (and subfolders) for files ending in");
-    console.log("    'automkv.yml'.  Any files found will have their destination folders");
-    console.log("    monitored for changes, and the edits in the file will run when any");
-    console.log("    updates have finished.");
-    console.log("  automkv run <target_file>");
-    console.log("    Apply the edits in a 'automkv.yml' file immediately to all target");
-    console.log("    files in the target folder");
-    Deno.exit(0);
-}
-
-export function readYaml(ymlPath: string): Batch[] {
-    const yml = (yaml.parse(Deno.readTextFileSync(ymlPath)) as { batch: Batch[] }).batch;
-    if (!yml) {
-        log.warning(`Invalid file: ${ymlPath}`);
-        log.warning("Expecting root element to be 'batch'");
-        log.warning("This file will be skipped.")
-        throw new Error("Improper YAML file");
-    }
-    // While functionally a map, we need to make it official
-    for (const batch of yml) {
-        batch.watch.files = new RegExp(batch.watch.files);
-        if (batch.edits)
-            for (const edit of batch.edits)
-                edit.set = new Map(Object.entries(edit.set));
-    }
-
-    return yml;
-}
+import {path} from "./deps.ts";
 
 async function getExeWindows(exe: string, decoder: TextDecoder) {
     return await Deno.run({
